[PLAY-1007] Added pemrissions module to test app
